pyproject.toml
[tool.poetry]
name = "SerPyTor"
version = "0.1.0"
description = "Distributed Production-level ML System for experimentation"
authors = ["Anuran Roy <76481787+anuran-roy@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
starlette = "^0.20.4"
polars = "^0.13.58"
uvicorn = "^0.18.2"
gunicorn = "^20.1.0"
pyspark = "^3.3.0"
scikit-learn = "^1.1.1"
rich = "^12.5.1"
pandas = "^1.4.3"
joblib = "^1.1.0"
tinydb = "^4.7.0"
PyYAML = "^6.0"
tqdm = "^4.64.0"
requests = "^2.28.1"
aiohttp = "^3.8.1"
aiodns = "^3.0.0"
Brotli = "^1.0.9"
cchardet = "^2.1.7"
upnpclient = "^1.0.3"
sentry-sdk = "^1.9.5"
networkx = "^2.8.6"
matplotlib = "^3.5.3"
plotly = "^5.10.0"
fugue = "^0.7.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
flake8 = "^5.0.0"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
